+++ _metadata.py
@@ -1,2 +1,2 @@
-__extension_version__ = "0.48.0"
+__extension_version__ = "0.52.0"
 __extension_name__ = "pytket-quantinuum"