+++ pytket/extensions/quantinuum/backends/quantinuum.py
@@ -80,6 +80,7 @@ from pytket.utils.outcomearray import OutcomeArray
 from pytket.wasm import WasmFileHandler
 
 from .api_wrappers import QuantinuumAPI, QuantinuumAPIError
+from .data import QuantinuumBackendData
 
 if TYPE_CHECKING:
     import matplotlib
@@ -118,6 +119,11 @@ _ADDITIONAL_GATES = {
     OpType.CopyBits,
     OpType.ClExpr,
     OpType.WASM,
+    OpType.RNGSeed,
+    OpType.RNGBound,
+    OpType.RNGIndex,
+    OpType.RNGNum,
+    OpType.JobShotNum,
 }
 
 _GATE_MAP = {
@@ -259,6 +265,12 @@ class _LocalEmulatorConfiguration:
     noisy_simulation: bool
 
 
+class BackendOfflineError(Exception):
+    """Raised when backend constructed with the `data` parameter is asked to make an
+    online API call.
+    """
+
+
 class QuantinuumBackend(Backend):
     """
     Interface to a Quantinuum device.
@@ -281,6 +293,7 @@ class QuantinuumBackend(Backend):
         machine_debug: bool = False,
         api_handler: QuantinuumAPI = DEFAULT_API_HANDLER,
         compilation_config: QuantinuumBackendCompilationConfig | None = None,
+        data: QuantinuumBackendData | None = None,
         **kwargs: QuumKwargTypes,
     ):
         """Construct a new Quantinuum backend.
@@ -295,6 +308,9 @@ class QuantinuumBackend(Backend):
             only support 'microsoft', which enables the microsoft Device Flow.
         :param api_handler: Instance of API handler, defaults to DEFAULT_API_HANDLER
         :param compilation_config: Optional compilation configuration
+        :param data: Data characterizing the backend. If this is not provided, the data
+            are retrieved online using the `device_name` provided. If it is provided,
+            then no online queries are made and no online submission is possible.
 
         Supported kwargs:
 
@@ -329,6 +345,8 @@ class QuantinuumBackend(Backend):
             self._compilation_config = QuantinuumBackendCompilationConfig()
         else:
             self._compilation_config = compilation_config
+
+        self._data = data
 
     @property
     def compilation_config(self) -> QuantinuumBackendCompilationConfig:
@@ -427,13 +445,29 @@ class QuantinuumBackend(Backend):
                 devices.append(cls._dict_to_backendinfo(d, local_emulator=True))
         return devices
 
-    def _retrieve_backendinfo(self, machine: str) -> BackendInfo:
-        infos = self.available_devices(api_handler=self.api_handler)
-        try:
-            info = next(entry for entry in infos if entry.device_name == machine)
-        except StopIteration:
-            raise DeviceNotAvailable(machine)  # noqa: B904
-        info.misc["options"] = self._process_circuits_options
+    def _retrieve_backendinfo(self) -> BackendInfo:
+        if self._data is None:
+            infos = self.available_devices(api_handler=self.api_handler)
+            try:
+                info = next(
+                    entry for entry in infos if entry.device_name == self._device_name
+                )
+            except StopIteration:
+                raise DeviceNotAvailable(self._device_name)  # noqa: B904
+            info.misc["options"] = self._process_circuits_options
+        else:
+            info = BackendInfo(
+                name=self.__class__.__name__,
+                device_name=self._device_name,
+                version=__extension_version__,
+                architecture=FullyConnected(self._data.n_qubits, "q"),
+                gate_set=set(self._data.gate_set),
+                n_cl_reg=self._data.n_cl_reg,
+                supports_fast_feedforward=True,
+                supports_midcircuit_measurement=True,
+                supports_reset=True,
+                misc={},
+            )
         return info
 
     @classmethod
@@ -511,6 +545,9 @@ class QuantinuumBackend(Backend):
             ``datetime.datetime`` objects.
         """
 
+        if self._data is not None:
+            raise BackendOfflineError("get_calendar() not available for this backend")
+
         if not isinstance(start_date, datetime.datetime) or not isinstance(
             end_date, datetime.datetime
         ):
@@ -579,6 +616,10 @@ class QuantinuumBackend(Backend):
         :return: A ``matplotlib.figure.Figure`` visualising the
             calendar for a user-specified calendar month.
         """
+
+        if self._data is not None:
+            raise BackendOfflineError("view_calendar() not available for this backend")
+
         if not MATPLOTLIB_IMPORT:
             raise ImportError(
                 "Matplotlib is not installed. Please run \
@@ -600,7 +641,7 @@ class QuantinuumBackend(Backend):
     @property
     def backend_info(self) -> BackendInfo | None:
         if self._backend_info is None and not self._MACHINE_DEBUG:
-            self._backend_info = self._retrieve_backendinfo(self._device_name)
+            self._backend_info = self._retrieve_backendinfo()
         return self._backend_info
 
     @cached_property
@@ -653,6 +694,8 @@ class QuantinuumBackend(Backend):
         """True if the backend is a local emulator, otherwise False"""
         if self._MACHINE_DEBUG:
             return False
+        if self._data is not None:
+            return self._data.local_emulator
         info = self.backend_info
         assert info is not None
         return bool(info.get_misc("system_type") == "local_emulator")
@@ -974,10 +1017,21 @@ class QuantinuumBackend(Backend):
         :return: ResultHandle for submitted job.
         """
 
+        if self._data is not None:
+            raise BackendOfflineError("submit_program() not available for this backend")
+
         if self.is_local_emulator:
             raise NotImplementedError(
                 "submit_program() not supported with local emulator"
             )
+
+        warnings.warn(
+            "Submission of programs to remote devices from pytket-quantinuum is "
+            "deprecated, and will not be possible after October 2025. Please use "
+            "qnexus ( https://docs.quantinuum.com/nexus/index.html ) instead.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
 
         lang_str = _language2str(language)
         if self.backend_info is not None:
@@ -1095,6 +1149,12 @@ class QuantinuumBackend(Backend):
         * `max_cost`: if set, the maximum amount in HQC to be spent on running the job.
           Ignored for local emulator.
         """
+
+        if self._data is not None and not self.is_local_emulator:
+            raise BackendOfflineError(
+                "process_circuits() not available for this backend"
+            )
+
         circuits = list(circuits)
         n_shots_list = Backend._get_n_shots_as_list(  # noqa: SLF001
             n_shots,
@@ -1465,7 +1525,7 @@ class QuantinuumBackend(Backend):
                         "Local emulator not available: \
 try installing with the `pecos` option."
                     )
-                from pytket_pecos import Emulator
+                from pytket_pecos import Emulator  # noqa: PLC0415
 
                 configuration = self._local_emulator_handles[handle]
                 # workaround for https://github.com/CQCL/pytket-quantinuum/issues/473
@@ -1523,8 +1583,10 @@ jobid is {jobid}"
     def cost_estimate(self, circuit: Circuit, n_shots: int) -> float | None:
         """Deprecated, use :py:meth:`cost`."""
 
-        warnings.warn(  # noqa: B028
-            "cost_estimate is deprecated, use cost instead", DeprecationWarning
+        warnings.warn(
+            "cost_estimate is deprecated, use cost instead",
+            DeprecationWarning,
+            stacklevel=2,
         )
 
         return self.cost(circuit, n_shots)
@@ -1560,6 +1622,10 @@ jobid is {jobid}"
         :raises ValueError: Circuit is not valid, needs to be compiled.
         :return: Cost in HQC to execute the shots.
         """
+
+        if self._data is not None:
+            raise BackendOfflineError("cost() not available for this backend")
+
         if not self.valid_circuit(circuit):
             raise ValueError(
                 "Circuit does not satisfy predicates of backend."  # noqa: ISC003
@@ -1747,6 +1813,10 @@ def _parse_status(response: dict) -> CircuitStatus:
             "queue-position",
             "cost",
             "error",
+            "cost-confidence",
+            "last-shot",
+            "qubits",
+            "priority",
         )
     }
     message = json.dumps(msgdict)