+++ docs/changelog.md
@@ -4,9 +4,23 @@
 
 # Changelog
 
-## Unreleased
+## 0.52.0 (July 2025)
+
+- Add optional argument `request_raw_results` to `QuantinuumAPI.retrieve_job_status`.
+
+## 0.51.0 (July 2025)
+
+- Update list of accepted op types for validity check.
+- Update pytket minimum version requirement to 2.7.0.
+
+## 0.50.0 (June 2025)
 
 - Update pytket-qir minimum version requirement to 0.24.1.
+- Emit deprecation warning from submit_program().
+- Add optional `data` argument to `QuantinuumBackend` constructor to allow
+  offline usage (as compilation engine or local emulator) when device data are
+  known.
+- Update pytket minimum version requirement to 2.6.0.
 
 ## 0.48.0 (May 2025)